Nixt can detect and respond to interactive prompts
Adds the public API methods on() and respond() for matching and
responding to interactive prompts.

See `test/prompt.test.js` for usage examples.

Inspired by https://github.com/jprichardson/node-suppose

Adds the 'string' module as a dependency for prompt matching.
Adds the 'prompt' module as a devDependency for the test fixture.